chore: rename segment.rs to state.rs
