+++ crates/vm/src/arch/mod.rs
@@ -11,9 +11,8 @@ mod integration_api;
 /// [RecordArena] trait definitions and implementations. Currently there are two concrete
 /// implementations: [MatrixRecordArena] and [DenseRecordArena].
 mod record_arena;
-/// Runtime execution and segmentation
-// TODO: rename this module
-pub mod segment;
+/// VM state definitions
+mod state;
 /// Top level [VmExecutor] and [VirtualMachine] constructor and API.
 pub mod vm;
 
@@ -31,5 +30,5 @@ pub use extensions::*;
 pub use integration_api::*;
 pub use openvm_instructions as instructions;
 pub use record_arena::*;
-pub use segment::*;
+pub use state::*;
 pub use vm::*;
